class2_prep/requirements.txt
pdfplumber
jieba
python-docx
wordcloud
